Redirect after creating a new repo
