+++ src/EmptyRepoPage/index.js
@@ -1,38 +1,35 @@
 import React, {Component} from 'react';
-import CreateRepoForm from './form';
+import EmptyRepoForm from './empty-repo-form';
 import UserInfoBlock from '../UserInfoBlock';
 import {connect} from "react-redux";
 import {userActions, userpageActions} from "../_actions";
-import {Redirect} from 'react-router-dom';
 
-class NewRepoPage extends Component {
+
+class EmptyRepoPage extends Component {
     componentDidMount() {
         this.props.getUserInfo();
     };
 
     render() {
-        console.log(this.props);
-        const {avatar, username, signOut, alert} = this.props;
-        const {repo} = this.props.repoCreation;
+        //location.pathname is /{username}/{repoName}/empty
+        const repoName = this.props.location.pathname.split('/')[2];
 
-        if (repo) {
-            return <Redirect to={`/${username}/${repo.name}/empty`} push={true}/>
-        }
-
+        const {avatar, username, signOut} = this.props;
         return (
             <div className="flex h-100">
                 <UserInfoBlock avatar={avatar} username={username} signOut={signOut}/>
                 <div className="container pt-5 w-100">
-                    <CreateRepoForm alert={alert}/>
+                    <EmptyRepoForm repoName={repoName}/>
                 </div>
             </div>
         );
     };
 }
 
+
 const mapStateToProps = state => {
     return {
-        ...state.userinfo, alert: state.alert, repoCreation: state.repoCreation
+        ...state.userinfo
     };
 };
 
@@ -47,6 +44,7 @@ const mapDispatchToProps = dispatch => {
     }
 };
 
-const ConnectedNewRepoPage = connect(mapStateToProps, mapDispatchToProps)(NewRepoPage);
 
-export default ConnectedNewRepoPage;
+const ConnectedEmptyRepoPage = connect(mapStateToProps, mapDispatchToProps)(EmptyRepoPage);
+
+export default ConnectedEmptyRepoPage;
