--- src/_components/empty-repo-page.js
@@ -1,13 +1,0 @@
-import React from 'react';
-import EmptyRepoForm from './empty-repo-form';
-import Page from './page';
-
-const EmptyRepository = props => {
-    return (
-        <Page>
-            <EmptyRepoForm/>
-        </Page>
-    );
-}
-
-export default EmptyRepository;
